Crate specs_hierarchy

Structs

Hierarchy
Scene graph type hierarchy.

HierarchySystem
System for maintaining a Hierarchy resource.

ParentData
Utility struct for the data needed by the Hierarchy maintain.

SubHierarchyIterator

Enums

HierarchyEvent
Hierarchy events.

Traits

Parent
Bound for the parent component of your crate. Your Parent component usually just contains the Entity that's the parent you're linking to.